data_prediction_service/requirements.txt
kafka-python==2.0.2
pyyaml==6.0
pandas==1.5.3